Support relative paths for logs, print help if no subcommand, quit pa… (#58)
* Support relative paths for logs, print help if no subcommand, quit parsing logs if it isnt json

* Remove services command

* Remove component update command
