+++ src/main/java/com/aws/greengrass/cli/CLI.java
@@ -8,7 +8,6 @@ package com.aws.greengrass.cli;
 import com.aws.greengrass.cli.commands.ComponentCommand;
 import com.aws.greengrass.cli.commands.DeploymentCommand;
 import com.aws.greengrass.cli.commands.Logs;
-import com.aws.greengrass.cli.commands.Service;
 import com.aws.greengrass.cli.module.AdapterModule;
 import com.aws.greengrass.cli.module.CommandsComponent;
 import com.aws.greengrass.cli.module.DaggerCommandsComponent;
@@ -26,11 +25,11 @@ import java.util.ResourceBundle;
  * Main entry point into the command line.
  */
 @Command(name = "cli",
-        subcommands = {HelpCommand.class, Service.class, ComponentCommand.class, DeploymentCommand.class, Logs.class},
+        subcommands = {HelpCommand.class, ComponentCommand.class, DeploymentCommand.class, Logs.class},
         resourceBundle = "com.aws.greengrass.cli.CLI_messages")
 public class CLI implements Runnable {
 
-    @CommandLine.Option(names = "--ggcRootPath")
+    @CommandLine.Option(names = "--ggcRootPath", description = "The path to the root directory of Greengrass")
     String ggcRootPath;
 
     @Spec
@@ -42,18 +41,15 @@ public class CLI implements Runnable {
         try {
             populateCommand(cli, args);
             exitCode = new CommandLine(cli, new CommandFactory(cli.createCommandComponent()))
-                    .setExecutionExceptionHandler(new CommandLine.IExecutionExceptionHandler() {
-                        @Override
-                        public int handleExecutionException(Exception e, CommandLine commandLine, CommandLine.ParseResult parseResult) throws Exception {
-                            if (e instanceof CommandLine.UnmatchedArgumentException
-                                    || e instanceof CommandLine.MissingParameterException
-                                    || e instanceof GenericCliIpcServerException) {
-                                System.out.println(commandLine.getColorScheme().errorText(e.getMessage()));
-                                commandLine.usage(System.out);
-                                return 0;
-                            } else {
-                                throw e;
-                            }
+                    .setExecutionExceptionHandler((e, commandLine, parseResult) -> {
+                        if (e instanceof CommandLine.UnmatchedArgumentException
+                                || e instanceof CommandLine.MissingParameterException
+                                || e instanceof GenericCliIpcServerException) {
+                            System.out.println(commandLine.getColorScheme().errorText(e.getMessage()));
+                            commandLine.usage(System.out);
+                            return 0;
+                        } else {
+                            throw e;
                         }
                     })
                     .execute(args);
