+++ src/main/java/com/aws/greengrass/cli/module/CommandsComponent.java
@@ -8,7 +8,6 @@ package com.aws.greengrass.cli.module;
 import com.aws.greengrass.cli.commands.ComponentCommand;
 import com.aws.greengrass.cli.commands.DeploymentCommand;
 import com.aws.greengrass.cli.commands.Logs;
-import com.aws.greengrass.cli.commands.Service;
 import dagger.Component;
 
 import javax.inject.Singleton;
@@ -24,6 +23,4 @@ public interface CommandsComponent {
     ComponentCommand component();
 
     DeploymentCommand deployment();
-
-    Service service();
 }
