+++ src/main/resources/com/aws/greengrass/cli/CLI_messages.properties
@@ -2,12 +2,9 @@ cli.usage.description=Greengrass command line interface%n
 cli.root=root directory of GGC
 cli.help.usage.header=Displays help information about a command.
 cli.health.help.usage.header=Displays help information about a command.
-cli.service.help.usage.header=Displays help information about a command.
 helpCommand.help=Show usage help for the help command and exit.
 helpCommand.command=The COMMAND to display the usage help message for.
-path=The path of a configuration key.
 exception.missing.command=A command is required.
-names=The names of services
 # Component
 cli.component.usage.description=List, get details, restart and stop components.
 cli.component.restart.usage.description=Restart component.
